Move kas::widget::view::ListMsg to kas::event::ChildMsg
